Port `search/pipeline` module to TS
