+++ src/search/add.ts
@@ -1,8 +1,9 @@
 import db, { VisitInteraction, PageAddRequest } from '.'
 import normalizeUrl from '../util/encode-url-for-id'
-import pipeline, { PipelineReq, transformUrl } from './pipeline'
+import pipeline, { transformUrl } from './pipeline'
 import { Page, FavIcon } from './models'
 import { getPage } from './util'
+import { PipelineReq } from './types'
 
 /**
  * Adds/updates a page + associated visit (pages never exist without either an assoc.
